move test helper functions
Move helper functions from partner_test.go to main_test.go
