+++ rsbe/main_test.go
@@ -2,6 +2,8 @@ package rsbe
 
 import (
 	"fmt"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"testing"
 )
@@ -46,4 +48,34 @@ func TestMain(m *testing.M) {
 	}
 
 	os.Exit(m.Run())
+}
+
+func setupMux(apiPath string, filePath string) (mux *http.ServeMux) {
+	mux = http.NewServeMux()
+	mux.HandleFunc(apiPath, func(w http.ResponseWriter, _ *http.Request) {
+		data, _ := os.ReadFile(filePath)
+		w.Write(data)
+	})
+
+	return mux
+}
+
+func setupTestServerClient(ts *httptest.Server) {
+	c := new(Config)
+	c.BaseURL = ts.URL
+	c.User = "foo"
+	c.Password = "bar"
+	c.AuthType = AuthTypeBasic
+
+	_ = ConfigureClient(c)
+}
+
+func setupLocalhostClient() {
+	//c, err := GetConfig("basic")
+	c, err := GetConfig("cookie")
+	if err != nil {
+		panic(err)
+	}
+
+	_ = ConfigureClient(c)
 }
